refactor opcodes to submodule
